feat(books): unify Genre picker — MudGenrePicker on Add / Edit too (PR 2)
Drew's testing-feedback 2026-05-12: "The genre presentation in full
edit (and therefore in add) is a horror story from the Cthulhu
mythos. I like the search in the edit work pop-up in the current
'view' page... 'Browse all Genres' expander that appears to not do
anything though, can probably be dropped."

The View page's WorkEditDialog already used MudGenrePicker (search
+ chips). Add and Edit pages used the older Bootstrap-era
GenrePicker — a wall of checkboxes for every parent + child genre.
This PR retires the Bootstrap picker and uses MudGenrePicker
everywhere.

Changes:

- MudGenrePicker.razor: dropped the collapsible "Browse all Genres"
  MudTreeView section (Drew flagged it as appearing to not do
  anything; uninspected on the rationale but dropped per his ask).
  Component is now chips + autocomplete only. Added LookupCandidates
  parameter (informational badges) + ApplyCandidatesAsync method
  (parent calls explicitly after lookup to fuzzy-match auto-select).
- MudGenrePickerViewModel: ported ApplyLookupCandidates (returns
  updated selection) and the FuzzyGenreMatch static (word-boundary
  regex) from the retired GenrePickerViewModel. BulkAddViewModel
  updated to call the new location.
- Add.razor + Edit.razor: swapped <GenrePicker> → <MudGenrePicker>,
  added "Genres" card wrapper for visual consistency, added the
  per-Work scope caption ("Genres apply to the work above" /
  "primary work above") so single-Work mode's implicit
  Book-equals-Work shortcut is labelled honestly.
- BookAddViewModel: dropped the GenrePickerViewModel? parameter from
  LookupAsync, ApplyCandidateAsync, and Reset. Picker now reads VM
  state via parameter binding and the page calls picker.Apply-
  CandidatesAsync explicitly after lookup. Cleaner contract.
- ProgramSetup: dropped AddTransient<GenrePickerViewModel>.
- Deleted GenrePicker.razor + GenrePickerViewModel.cs (no remaining
  consumers).

Tests:

- Deleted GenrePickerViewModelTests.cs.
- Ported the 11 FuzzyGenreMatch theory cases to MudGenrePicker-
  ViewModelTests so the matcher's word-boundary protection stays
  covered (Romance vs Romanticism, Science vs Science Fiction, etc).
- BookAddViewModelTests: dropped CreateGenrePicker() helper +
  removed it from 8 LookupAsync / ApplyCandidateAsync call sites.

43 of 43 BookAdd + MudGenrePicker tests pass. Visual verification of
the new "Genres" card on Add and Edit pages welcome on local before
push — should read as chips + autocomplete, no tree, with a small
per-Work caption above.
