+++ BookTracker.Web/Components/Pages/Books/Edit.razor
@@ -49,11 +49,14 @@ else
             }
 
             <div class="col-12">
-                <GenrePicker @ref="genrePicker"
-                             SelectedGenreIds="VM.SelectedGenreIds"
-                             SelectedGenreIdsChanged="ids => VM.SelectedGenreIds = ids"
-                             LookupCandidates="[]" />
-                <div class="form-text mt-1">Genres apply to the primary work above.</div>
+                <div class="card">
+                    <div class="card-header bg-white"><h2 class="h5 mb-0">Genres</h2></div>
+                    <div class="card-body">
+                        <div class="form-text mb-2">Genres apply to the primary work above.</div>
+                        <MudGenrePicker SelectedGenreIds="VM.SelectedGenreIds"
+                                        SelectedGenreIdsChanged="ids => VM.SelectedGenreIds = ids" />
+                    </div>
+                </div>
             </div>
 
             <div class="col-12">
@@ -368,7 +371,6 @@ else
     [Parameter]
     public int BookId { get; set; }
 
-    private GenrePicker? genrePicker;
 
     protected override async Task OnInitializedAsync() => await VM.InitializeAsync(BookId);
