--- BookTracker.Web/Components/Shared/GenrePicker.razor
@@ -1,69 +1,0 @@
-@inject GenrePickerViewModel VM
-
-<div class="card">
-    <div class="card-header bg-white"><h2 class="h5 mb-0">Genres</h2></div>
-    <div class="card-body">
-        <div class="row g-3">
-            @foreach (var top in VM.TopLevelGenres)
-            {
-                <div class="col-md-6 col-lg-4">
-                    <div class="form-check">
-                        <input class="form-check-input" type="checkbox" id="@($"genre-{top.Id}")"
-                               checked="@VM.SelectedGenreIds.Contains(top.Id)"
-                               @onchange="e => OnToggle(top.Id, (bool)e.Value!)" />
-                        <label class="form-check-label fw-semibold" for="@($"genre-{top.Id}")">@top.Name</label>
-                    </div>
-                    @foreach (var child in top.Children.OrderBy(c => c.Name))
-                    {
-                        <div class="form-check ms-4">
-                            <input class="form-check-input" type="checkbox" id="@($"genre-{child.Id}")"
-                                   checked="@VM.SelectedGenreIds.Contains(child.Id)"
-                                   @onchange="e => OnToggle(child.Id, (bool)e.Value!)" />
-                            <label class="form-check-label" for="@($"genre-{child.Id}")">@child.Name</label>
-                        </div>
-                    }
-                </div>
-            }
-        </div>
-        @if (VM.LookupCandidates.Count > 0)
-        {
-            <div class="mt-3 pt-3 border-top">
-                <small class="text-muted d-block mb-1">Lookup returned (fuzzy-matched against the list above):</small>
-                @foreach (var c in VM.LookupCandidates)
-                {
-                    <span class="badge bg-light text-dark border me-1 mb-1">@c</span>
-                }
-            </div>
-        }
-    </div>
-</div>
-
-@code {
-    [Parameter, EditorRequired]
-    public List<int> SelectedGenreIds { get; set; } = [];
-
-    [Parameter]
-    public EventCallback<List<int>> SelectedGenreIdsChanged { get; set; }
-
-    [Parameter]
-    public List<string> LookupCandidates { get; set; } = [];
-
-    /// <summary>
-    /// Exposes the underlying ViewModel so parent components can call
-    /// ApplyLookupCandidates and access genre data directly.
-    /// </summary>
-    public GenrePickerViewModel ViewModel => VM;
-
-    protected override async Task OnInitializedAsync()
-    {
-        VM.SelectedGenreIds = SelectedGenreIds;
-        VM.LookupCandidates = LookupCandidates;
-        await VM.InitializeAsync();
-    }
-
-    private async Task OnToggle(int id, bool isChecked)
-    {
-        VM.ToggleGenre(id, isChecked);
-        await SelectedGenreIdsChanged.InvokeAsync(VM.SelectedGenreIds);
-    }
-}
