+++ BookTracker.Web/ProgramSetup.cs
@@ -153,7 +153,6 @@ public static class ProgramSetup
         builder.Services.AddTransient<BookFormViewModel>();
         builder.Services.AddTransient<EditionFormViewModel>();
         builder.Services.AddTransient<CopyFormViewModel>();
-        builder.Services.AddTransient<GenrePickerViewModel>();
         builder.Services.AddTransient<MudGenrePickerViewModel>();
         builder.Services.AddTransient<BookListViewModel>();
         builder.Services.AddTransient<BookAddViewModel>();
